fix(resolver): also disable PluginImporter when stale-flat sweep hits a file lock
The user's CS1704 reproduction: stale McpPlugin.6.2.0.dll on disk alongside
the canonical 6.2.1, both with assembly name "McpPlugin" — Unity loads one,
locks it, and RemoveStaleVersionDllsByStem's TryDeleteFile silently logged
and gave up. The sweep then reported success, the file survived, and the
duplicate-assembly error never cleared.

Pull the locked-file recovery path out of NuGetLegacyMigration (where it
already existed as a private helper) into a shared
NuGetPluginConfigurator.DisableImporter, and call it from the stale-flat
sweep on IOException / UnauthorizedAccessException. The next domain reload
unloads the disabled DLL, the file unlocks, and the next sweep pass deletes
it. Two-cycle recovery instead of "stuck forever".
